requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
pyjwt>=2.8.0
